% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csor_shiny.R
\name{csor_shiny}
\alias{csor_shiny}
\title{Shiny App to show the plots from the csor package}
\usage{
csor_shiny()
}
\value{
Runs the shiny app
}
\description{
Shiny App to show the plots from the csor package
}
\examples{
\dontrun{
csor_shiny()
}
}

