% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.csor.R
\name{fit}
\alias{fit}
\title{Fit a smoothing spline to one of the GDP time series}
\usage{
fit(obj, SecCode = 0, SeaAdj = FALSE)
}
\arguments{
\item{obj}{An object of class \code{csor} from \code{\link{loadGDP}}}

\item{SecCode}{The sector code for the sector to fit}

\item{SeaAdj}{An indicator of whether we are fitting to the seasonally adjusted (TRUE) or the non-seasonally adjusted (FALSE) data}
}
\value{
Returns a list of class \code{smooth.spline}
}
\description{
Fit a smoothing spline to one of the GDP time series
}
\examples{
data = loadGDP()
fit1 = fit(data)
fit2 = fit(data, SecCode = 4, SeaAdj = TRUE)
}
\seealso{
\code{\link{loadGDP}}, \code{\link{plot.csor}}, \code{\link{sectorList}}, \code{\link{predict.csor}}, \code{\link{smooth.spline}}
}

