% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGDP.R
\name{loadGDP}
\alias{loadGDP}
\title{Load GDP Data from the CSO}
\usage{
loadGDP()
}
\value{
A csor object containing the latest available GDP data from the CSO by sector, seasonally adjusted and not seasonally adjusted
}
\description{
Load GDP Data from the CSO
}
\examples{
data = loadGDP()
}
\seealso{
\code{\link{sectorList}}, \code{\link{plot.csor}}, \code{\link{fit.csor}}, \code{\link{predict.csor}}
}

