% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.csor.R
\name{predict.csor}
\alias{predict.csor}
\title{Predict one quarter ahead for each of the GDP time series}
\usage{
\method{predict}{csor}(object, ...)
}
\arguments{
\item{object}{An object of class \code{csor} from \code{\link{loadGDP}}}

\item{...}{Additional arguments passed to or from other methods}
}
\value{
A data frame wih a prediction for the next quarters GDP for each combination of sector and whether the data is seasonally adjusted or not.
}
\description{
Predict one quarter ahead for each of the GDP time series
}
\examples{
data = loadGDP()
pred = predict(data)
}
\seealso{
\code{\link{loadGDP}}, \code{\link{plot.csor}}, \code{\link{sectorList}}, \code{\link{fit}}, \code{\link{smooth.spline}}
}

