% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sectorList.R
\name{sectorList}
\alias{sectorList}
\title{Return the list of sectors with their codes}
\usage{
sectorList(data)
}
\arguments{
\item{data}{Output from the \code{\link{loadGDP}} function}
}
\value{
Nothing: prints the list
}
\description{
Return the list of sectors with their codes
}
\examples{
data = loadGDP()
sectorList(data)
}
\seealso{
\code{\link{loadGDP}}, \code{\link{plot.csor}}, \code{\link{fit}}, \code{\link{predict.csor}}
}

